python/auth/requirements.txt
click==6.7
Flask==0.12.2
Flask-JSON==0.3.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
marshmallow==2.15.0
pika==0.11.2
PyJWT==1.4.2
PyYAML==3.12
Werkzeug==0.14.1
